Function to find best model by information criteria, by model type

Function to find best model by information criteria, by model type

Usage

findBestModelFunc(
  obsdatval,
  modType,
  requiredVarNames,
  allVarNames,
  complexModel,
  randomEffects = NULL,
  useParallel,
  selectCriteria,
  varExclude,
  printOutput = FALSE,
  catchType = NULL,
  common = NULL,
  dirname = NULL,
  run = NULL,
  modelScenario = NULL,
  shortName = NULL
)

Arguments

obsdatval

Value

modType

Value

requiredVarNames

Value

allVarNames

Value

complexModel

Value

randomEffects

Value

useParallel

Value

selectCriteria

Value

varExclude

Value

printOutput

Value

catchType

Value

common

Value

dirname

Value

run

Value

modelScenario

Value

shortName

Value