loadOutputs

Reads in all the R objects created during runs of the bycatchEstimator for further analysis

Usage

loadOutputs(
  baseDir = getwd(),
  runName,
  runDate = Sys.Date(),
  loadDesign = TRUE,
  designScenario = NULL,
  loadModel = TRUE,
  modelScenario = NULL
)

Arguments

baseDir

The base directory for the runs, same as bycatchSetup.

runName

The run name, same as bycatchSetup.

runDate

The date when the model was run. Defaults to current date, but can be set to read in models previously run.

loadDesign

TRUE to read in design-based estimator results.

designScenario

Value of designScenario from original run.

loadModel

TRUE to read in model-based estimator results.

modelScenario

Value of designScenario from original run.